docs: finishing docs for spring 3 + Postman Json

+++ src/main/java/br/com/intellistocks/api/service/TypeProductService.java
@@ -4,10 +4,12 @@ import br.com.intellistocks.api.models.product.TypeProduct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import br.com.intellistocks.api.repository.TypeProductRepository;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class TypeProductService {
@@ -35,11 +37,15 @@ public class TypeProductService {
     }
 
     public TypeProduct editTypeProduct(Long id, TypeProduct typeProduct) {
+        if (!typeProductRepository.existsById(id)) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "TypeProduct " +
+                "not found! with id: " + id);
         typeProduct.setId(id);
         return typeProductRepository.save(typeProduct);
     }
 
     public void deleteTypeProduct(Long id) {
+        if (!typeProductRepository.existsById(id)) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "TypeProduct " +
+                "not found! with id: " + id);
         typeProductRepository.deleteById(id);
     }
 }
